impr: split direction helper service into three classes (FPLA-NA) (#778)
