+++ service/src/main/java/uk/gov/hmcts/reform/fpl/controllers/ComplyOnBehalfController.java
@@ -2,6 +2,7 @@ package uk.gov.hmcts.reform.fpl.controllers;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import io.swagger.annotations.Api;
+import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -17,8 +18,10 @@ import uk.gov.hmcts.reform.fpl.model.CaseData;
 import uk.gov.hmcts.reform.fpl.model.Direction;
 import uk.gov.hmcts.reform.fpl.model.Others;
 import uk.gov.hmcts.reform.fpl.model.common.Element;
-import uk.gov.hmcts.reform.fpl.service.DirectionHelperService;
+import uk.gov.hmcts.reform.fpl.service.CommonDirectionService;
 import uk.gov.hmcts.reform.fpl.service.OthersService;
+import uk.gov.hmcts.reform.fpl.service.PrepareDirectionsForDataStoreService;
+import uk.gov.hmcts.reform.fpl.service.PrepareDirectionsForUsersService;
 import uk.gov.hmcts.reform.fpl.service.RespondentService;
 
 import java.util.List;
@@ -30,22 +33,14 @@ import static net.logstash.logback.encoder.org.apache.commons.lang3.ObjectUtils.
 @Api
 @RestController
 @RequestMapping("/callback/comply-on-behalf")
+@RequiredArgsConstructor(onConstructor = @__(@Autowired))
 public class ComplyOnBehalfController {
     private final ObjectMapper mapper;
-    private final DirectionHelperService directionHelperService;
+    private final CommonDirectionService commonDirectionService;
+    private final PrepareDirectionsForDataStoreService prepareDirectionsForDataStoreService;
+    private final PrepareDirectionsForUsersService prepareDirectionsForUsersService;
     private final RespondentService respondentService;
     private final OthersService othersService;
-
-    @Autowired
-    public ComplyOnBehalfController(ObjectMapper mapper,
-                                    DirectionHelperService directionHelperService,
-                                    RespondentService respondentService,
-                                    OthersService othersService) {
-        this.mapper = mapper;
-        this.directionHelperService = directionHelperService;
-        this.respondentService = respondentService;
-        this.othersService = othersService;
-    }
 
     //TODO: filter responses with different userName in aboutToStart. Code below makes the assumption that only
     // the same responder will be able edit a response. Currently any solicitor can amend a response but the
@@ -55,14 +50,14 @@ public class ComplyOnBehalfController {
         CaseDetails caseDetails = callbackrequest.getCaseDetails();
         CaseData caseData = mapper.convertValue(caseDetails.getData(), CaseData.class);
 
-        List<Element<Direction>> directionsToComplyWith = directionHelperService.getDirectionsToComplyWith(caseData);
+        List<Element<Direction>> directionsToComplyWith = commonDirectionService.getDirectionsToComplyWith(caseData);
 
         Map<DirectionAssignee, List<Element<Direction>>> sortedDirections =
-            directionHelperService.sortDirectionsByAssignee(directionsToComplyWith);
+            commonDirectionService.sortDirectionsByAssignee(directionsToComplyWith);
 
-        directionHelperService.addEmptyDirectionsForAssigneeNotInMap(sortedDirections);
+        commonDirectionService.addEmptyDirectionsForAssigneeNotInMap(sortedDirections);
 
-        directionHelperService.addDirectionsToCaseDetails(
+        prepareDirectionsForUsersService.addDirectionsToCaseDetails(
             caseDetails, sortedDirections, ComplyOnBehalfEvent.valueOf(callbackrequest.getEventId()));
 
         caseDetails.getData().put("respondents_label", getRespondentsLabel(caseData));
@@ -80,7 +75,7 @@ public class ComplyOnBehalfController {
         CaseDetails caseDetails = callbackrequest.getCaseDetails();
         CaseData caseData = mapper.convertValue(caseDetails.getData(), CaseData.class);
 
-        directionHelperService.addComplyOnBehalfResponsesToDirectionsInOrder(
+        prepareDirectionsForDataStoreService.addComplyOnBehalfResponsesToDirectionsInOrder(
             caseData, ComplyOnBehalfEvent.valueOf(callbackrequest.getEventId()), authorisation);
 
         //TODO: new service for sdo vs cmo in placing directions
