+++ service/src/main/java/uk/gov/hmcts/reform/fpl/controllers/DraftOrdersController.java
@@ -24,9 +24,10 @@ import uk.gov.hmcts.reform.fpl.model.common.DocmosisDocument;
 import uk.gov.hmcts.reform.fpl.model.common.DocumentReference;
 import uk.gov.hmcts.reform.fpl.model.common.Element;
 import uk.gov.hmcts.reform.fpl.service.CaseDataExtractionService;
-import uk.gov.hmcts.reform.fpl.service.DirectionHelperService;
+import uk.gov.hmcts.reform.fpl.service.CommonDirectionService;
 import uk.gov.hmcts.reform.fpl.service.DocmosisDocumentGeneratorService;
 import uk.gov.hmcts.reform.fpl.service.OrdersLookupService;
+import uk.gov.hmcts.reform.fpl.service.PrepareDirectionsForDataStoreService;
 import uk.gov.hmcts.reform.fpl.service.UploadDocumentService;
 import uk.gov.hmcts.reform.fpl.service.ccd.CoreCaseDataService;
 
@@ -48,28 +49,31 @@ public class DraftOrdersController {
     private final DocmosisDocumentGeneratorService docmosisService;
     private final UploadDocumentService uploadDocumentService;
     private final CaseDataExtractionService caseDataExtractionService;
-    private final DirectionHelperService directionHelperService;
+    private final CommonDirectionService commonDirectionService;
     private final OrdersLookupService ordersLookupService;
     private final CoreCaseDataService coreCaseDataService;
     private final ApplicationEventPublisher applicationEventPublisher;
+    private final PrepareDirectionsForDataStoreService prepareDirectionsForDataStoreService;
 
     @Autowired
     public DraftOrdersController(ObjectMapper mapper,
                                  DocmosisDocumentGeneratorService docmosisService,
                                  UploadDocumentService uploadDocumentService,
                                  CaseDataExtractionService caseDataExtractionService,
-                                 DirectionHelperService directionHelperService,
+                                 CommonDirectionService commonDirectionService,
                                  OrdersLookupService ordersLookupService,
                                  CoreCaseDataService coreCaseDataService,
-                                 ApplicationEventPublisher applicationEventPublisher) {
+                                 ApplicationEventPublisher applicationEventPublisher,
+                                 PrepareDirectionsForDataStoreService prepareDirectionsForDataStoreService) {
         this.mapper = mapper;
         this.docmosisService = docmosisService;
         this.uploadDocumentService = uploadDocumentService;
         this.caseDataExtractionService = caseDataExtractionService;
-        this.directionHelperService = directionHelperService;
+        this.commonDirectionService = commonDirectionService;
         this.ordersLookupService = ordersLookupService;
         this.coreCaseDataService = coreCaseDataService;
         this.applicationEventPublisher = applicationEventPublisher;
+        this.prepareDirectionsForDataStoreService = prepareDirectionsForDataStoreService;
     }
 
     @PostMapping("/about-to-start")
@@ -92,10 +96,10 @@ public class DraftOrdersController {
     }
 
     private Map<DirectionAssignee, List<Element<Direction>>> sortDirectionsByAssignee(CaseData caseData) {
-        List<Element<Direction>> nonCustomDirections = directionHelperService
+        List<Element<Direction>> nonCustomDirections = commonDirectionService
             .removeCustomDirections(caseData.getStandardDirectionOrder().getDirections());
 
-        return directionHelperService.sortDirectionsByAssignee(nonCustomDirections);
+        return commonDirectionService.sortDirectionsByAssignee(nonCustomDirections);
     }
 
     @PostMapping("/mid-event")
@@ -108,12 +112,12 @@ public class DraftOrdersController {
 
         CaseData updated = caseData.toBuilder()
             .standardDirectionOrder(Order.builder()
-                .directions(directionHelperService.combineAllDirections(caseData))
+                .directions(commonDirectionService.combineAllDirections(caseData))
                 .judgeAndLegalAdvisor(caseData.getJudgeAndLegalAdvisor())
                 .build())
             .build();
 
-        directionHelperService.persistHiddenDirectionValues(
+        prepareDirectionsForDataStoreService.persistHiddenDirectionValues(
             getConfigDirectionsWithHiddenValues(), updated.getStandardDirectionOrder().getDirections());
 
         Document document = getDocument(
@@ -147,13 +151,13 @@ public class DraftOrdersController {
 
         CaseData updated = caseData.toBuilder()
             .standardDirectionOrder(Order.builder()
-                .directions(directionHelperService.combineAllDirections(caseData))
+                .directions(commonDirectionService.combineAllDirections(caseData))
                 .orderStatus(caseData.getStandardDirectionOrder().getOrderStatus())
                 .judgeAndLegalAdvisor(caseData.getJudgeAndLegalAdvisor())
                 .build())
             .build();
 
-        directionHelperService.persistHiddenDirectionValues(
+        prepareDirectionsForDataStoreService.persistHiddenDirectionValues(
             getConfigDirectionsWithHiddenValues(), updated.getStandardDirectionOrder().getDirections());
 
         Document document = getDocument(
@@ -207,7 +211,7 @@ public class DraftOrdersController {
         // constructDirectionForCCD requires LocalDateTime, but this value is not used in what is returned
         return ordersLookupService.getStandardDirectionOrder().getDirections()
             .stream()
-            .map(direction -> directionHelperService.constructDirectionForCCD(direction, LocalDateTime.now()))
+            .map(direction -> commonDirectionService.constructDirectionForCCD(direction, LocalDateTime.now()))
             .collect(Collectors.toList());
     }
