+++ service/src/main/java/uk/gov/hmcts/reform/fpl/model/DirectionResponse.java
@@ -15,8 +15,8 @@ import java.util.UUID;
 @Data
 @Builder(toBuilder = true)
 public class DirectionResponse {
-    private final UUID directionId;
-    private final DirectionAssignee assignee;
+    private UUID directionId;
+    private DirectionAssignee assignee;
     private final String responder;
     private final String complied;
     private final String documentDetails;
