+++ service/src/main/java/uk/gov/hmcts/reform/fpl/service/DraftCMOService.java
@@ -44,7 +44,7 @@ import static uk.gov.hmcts.reform.fpl.enums.DirectionAssignee.PARENTS_AND_RESPON
 @RequiredArgsConstructor(onConstructor = @__(@Autowired))
 public class DraftCMOService {
     private final DateFormatterService dateFormatterService;
-    private final DirectionHelperService directionHelperService;
+    private final CommonDirectionService commonDirectionService;
 
     public Map<String, Object> extractIndividualCaseManagementOrderObjects(
         CaseManagementOrder caseManagementOrder,
@@ -111,7 +111,7 @@ public class DraftCMOService {
 
     public void prepareCustomDirections(CaseDetails caseDetails, CaseManagementOrder order) {
         if (!isNull(order)) {
-            directionHelperService.sortDirectionsByAssignee(order.getDirections())
+            commonDirectionService.sortDirectionsByAssignee(order.getDirections())
                 .forEach((key, value) -> caseDetails.getData().put(key.getValue(), value));
         } else {
             removeExistingCustomDirections(caseDetails);
@@ -150,22 +150,22 @@ public class DraftCMOService {
     private List<Element<Direction>> combineAllDirectionsForCmo(CaseData caseData) {
         List<Element<Direction>> directions = new ArrayList<>();
 
-        directions.addAll(directionHelperService.assignCustomDirections(caseData.getAllPartiesCustomCMO(),
+        directions.addAll(commonDirectionService.assignCustomDirections(caseData.getAllPartiesCustomCMO(),
             ALL_PARTIES));
 
-        directions.addAll(directionHelperService.assignCustomDirections(caseData.getLocalAuthorityDirectionsCustomCMO(),
+        directions.addAll(commonDirectionService.assignCustomDirections(caseData.getLocalAuthorityDirectionsCustomCMO(),
             LOCAL_AUTHORITY));
 
-        directions.addAll(orderByParentsAndRespondentAssignee(directionHelperService.assignCustomDirections(
+        directions.addAll(orderByParentsAndRespondentAssignee(commonDirectionService.assignCustomDirections(
             caseData.getRespondentDirectionsCustomCMO(), PARENTS_AND_RESPONDENTS)));
 
-        directions.addAll(directionHelperService.assignCustomDirections(caseData.getCafcassDirectionsCustomCMO(),
+        directions.addAll(commonDirectionService.assignCustomDirections(caseData.getCafcassDirectionsCustomCMO(),
             CAFCASS));
 
-        directions.addAll(orderByOtherPartiesAssignee(directionHelperService.assignCustomDirections(
+        directions.addAll(orderByOtherPartiesAssignee(commonDirectionService.assignCustomDirections(
             caseData.getOtherPartiesDirectionsCustomCMO(), OTHERS)));
 
-        directions.addAll(directionHelperService.assignCustomDirections(caseData.getCourtDirectionsCustomCMO(), COURT));
+        directions.addAll(commonDirectionService.assignCustomDirections(caseData.getCourtDirectionsCustomCMO(), COURT));
 
         return directions;
     }
